hopefully improve readability, move files to /components/sidebar
