+++ apps/georgeai-webapp/src/components/sidebar/sidebar-create-new-item-button.tsx.tsx
@@ -1,0 +1,25 @@
+import { PlusIcon } from '../../icons/plus-icon'
+
+interface SidebarCreateNewItemButtonProps {
+  newItemDialogRef: React.RefObject<HTMLDialogElement | null>
+  createNewItemTooltip: string
+  hoverClass: string
+}
+
+export const SidebarCreateNewItemButton = ({
+  newItemDialogRef: newDialogRef,
+  createNewItemTooltip,
+  hoverClass,
+}: SidebarCreateNewItemButtonProps) => (
+  <button
+    type="button"
+    onClick={(e) => {
+      e.stopPropagation()
+      newDialogRef.current?.showModal()
+    }}
+    className={`tooltip btn tooltip-left right-2 z-20 ml-auto btn-circle shrink-0 opacity-0 btn-ghost btn-xs [&::before]:text-xs ${hoverClass}`}
+    data-tip={createNewItemTooltip}
+  >
+    <PlusIcon />
+  </button>
+)
