+++ apps/georgeai-webapp/src/components/sidebar/sidebar-divider.tsx
@@ -1,0 +1,7 @@
+export function SidebarDivider() {
+  return (
+    <li className="pointer-events-none">
+      <div className="divider my-0 px-2" />
+    </li>
+  )
+}
