+++ apps/georgeai-webapp/src/routes/__root.tsx
@@ -8,7 +8,7 @@ import { AuthProvider } from '../auth/auth'
 import { getUserQueryOptions } from '../auth/get-user'
 import BottomNavigationMobile from '../components/bottom-navigation-mobile'
 import { GeorgeToaster } from '../components/georgeToaster'
-import { SidebarLayout } from '../components/sidebar-layout'
+import { SidebarLayout } from '../components/sidebar/sidebar-layout'
 import { getThemeQueryOptions } from '../hooks/use-theme'
 import { getLanguageQueryOptions, useLanguage } from '../i18n/use-language-hook'
 import appCss from '../index.css?url'
